codebase-query-contextualiser/function-calling/requirements.txt
aiohttp==3.9.3
langchain-core==0.1.32
python-dotenv==1.0.1 